Rationalise logic
All `install.sh`, `build.setup.sh` and `build.run.sh` sub-scripts can be OS independent,
or placed in the appropriate `linux` or `windows` sub-directory.

This applies to the root `travis` directory as well as the arch ones.

Sub-scripts which are always sourced from the main script do not need shebang, etc.
so these have been removed; it also clarifies that the script is not to be run stand-alone.